Added initial UI.
Index.html the main page now has it's initial logged out ui. The sign in and signup do not work yet. Auth.js not checks for the authentication state change and calls a function in index.js. Index.js setsup modals, as well as controls what components are seen when logged in or out.