Began work on chat file upload and emote recognition. Added some quotes.
Chat - Added begining of  emote recognition. Still need to program emote replacement.

Added begining of file upload. File is sent to server and checked. Was waiting on firebase admin module to be installed. Firebase admin is needed to send the message from the server, because the virus scan report is lost when sent to the client from server.